Alerting: Fix unique violation when updating rule group with title chains/cycles
The uniqueness constraint for titles within an org+folder is enforced on every update within a transaction instead of on commit (deferred constraint). This can cause problems when updating multiple rule titles at the same time:

1) Chains: Assuming Rule A, B, and C exist, and we have an update chain of RuleA -> RuleB -> RuleC -> RuleNew. We need to update in reverse order to avoid a constraint violation; RuleC -> RuleNew, RuleB -> RuleC, RuleA -> RuleB.
2) Cycles: Assuming Rule A, B, and C exist, and we have an update chain of RuleA -> RuleB -> RuleC -> RuleA. We need to break the cycle by adding a temporary update of RuleA -> RuleTemp, then we can update in reverse order; RuleC -> RuleA, RuleB -> RuleC, RuleA -> RuleB.

In addition, we make sure diffs are executed in the following order: DELETES, UPDATES, INSERTS.

The alternatives to this added complexity are:

1) On update delete entire rule group and reinsert. This is simple and straightforward, but requires removing our optimistic locking strategy for pessimistic.
2) Remove the title uniqueness constraint from the table. Then we can either enforce in application code or not enforce it at all. Not enforcing it could cause users to mistakenly group notifications for alerts of the same name.

Note: Deferred constraints are not really a viable option here as SQLite and MySQL do not support it at the moment.