Replace exact solution with simpler inexact implementation
The exact solution to this is complex and requires determining directed
paths and cycles in the update graph, adding in temporary updates to break
cycles, and then executing the updates in reverse topological order.

Instead, we choose a simpler solution that works
in all cases but might perform more updates than necessary. This simpler
solution makes a determination of whether an intermediate collision could
occur and if so, adds a temporary title on all updated rules to break any
cycles and remove the need for specific ordering.